dip code cleanup use size_t type instead of unsigned int type for memory size

+++ bld/dig/c/dipman.c
@@ -783,7 +783,7 @@ const char *ImageDIP( mod_handle mh )
 /*
  * Module Information
  */
-unsigned ModName( mod_handle mh, char *buff, unsigned buff_size )
+size_t ModName( mod_handle mh, char *buff, size_t buff_size )
 {
     image_handle        *ih;
 
@@ -1050,16 +1050,14 @@ int TypeCmp( type_handle *th1, type_handle *th2 )
     return( ih->dip->type_cmp( IMP_HDL( ih, image ), IMP_HDL( th1, type ), IMP_HDL( th2, type ) ) );
 }
 
-unsigned TypeName( type_handle *th, unsigned num, symbol_type *tag,
-                        char *buff, unsigned buff_size )
+size_t TypeName( type_handle *th, unsigned num, symbol_type *tag, char *buff, size_t buff_size )
 {
     image_handle        *ih;
 
     ih = II2IH( th->ii );
     if( ih == NULL )
         return( 0 );
-    return( ih->dip->type_name( IMP_HDL( ih, image ),
-                IMP_HDL( th, type ), num, tag, buff, buff_size ) );
+    return( ih->dip->type_name( IMP_HDL( ih, image ), IMP_HDL( th, type ), num, tag, buff, buff_size ) );
 }
 
 /*
@@ -1074,16 +1072,14 @@ mod_handle SymMod( sym_handle *sh )
 }
 
 //NYI: needs to do something for expression names
-unsigned SymName( sym_handle *sh, location_context *lc, symbol_name sn,
-                        char *buff, unsigned buff_size )
+size_t SymName( sym_handle *sh, location_context *lc, symbol_name sn, char *buff, size_t buff_size )
 {
     image_handle        *ih;
 
     ih = II2IH( sh->ii );
     if( ih == NULL )
         return( 0 );
-    return( ih->dip->sym_name( IMP_HDL( ih, image ),
-                IMP_HDL( sh, sym ), lc, sn, buff, buff_size ) );
+    return( ih->dip->sym_name( IMP_HDL( ih, image ), IMP_HDL( sh, sym ), lc, sn, buff, buff_size ) );
 }
 
 dip_status SymType( sym_handle *sh, type_handle *th )
@@ -1235,7 +1231,7 @@ mod_handle CueMod( cue_handle *ch )
     return( MK_MH( ih->ii, ih->dip->cue_mod( IMP_HDL( ih, image ), IMP_HDL( ch, cue ) ) ) );
 }
 
-unsigned CueFile( cue_handle *ch, char *buff, unsigned buff_size )
+size_t CueFile( cue_handle *ch, char *buff, size_t buff_size )
 {
     image_handle        *ih;
