+++ bld/dip/javavm/h/jvmdip.h
@@ -119,7 +119,7 @@ unsigned        GetU8( ji_ptr off );
 unsigned        GetU16( ji_ptr off );
 unsigned long   GetU32( ji_ptr off );
 
-unsigned        NameCopy( char *buff, const char *src, unsigned buff_size, unsigned len );
+size_t          NameCopy( char *buff, const char *src, size_t buff_size, size_t len );
 void            NormalizeClassName( char *p, unsigned len );
