+++ bld/dip/watcom/h/watmod.h
@@ -39,12 +39,12 @@ extern unsigned         ModInfoSplit( info_block *blk, section_info *inf );
 extern unsigned         ModOff2Idx( section_info *inf, unsigned off );
 extern walk_result      MyWalkModList( imp_image_handle *ii, INT_MOD_WKR *wk,
                                 void *d );
-extern unsigned         PrimaryCueFile( imp_image_handle *ii, imp_cue_handle *ic,
-                                char *buff, unsigned buff_size );
+extern size_t           PrimaryCueFile( imp_image_handle *ii, imp_cue_handle *ic,
+                                char *buff, size_t buff_size );
 extern walk_result DIGENTRY DIPImpWalkModList( imp_image_handle *ii, IMP_MOD_WKR *wk,
                                 void *d );
 extern char *DIGENTRY       DIPImpModSrcLang( imp_image_handle *ii, imp_mod_handle im );
 extern dip_status DIGENTRY  DIPImpModInfo( imp_image_handle *ii, imp_mod_handle im,
                                 handle_kind hk );
-extern unsigned DIGENTRY    DIPImpModName( imp_image_handle *ii, imp_mod_handle im,
-                                char *buff, unsigned buff_size );
+extern size_t DIGENTRY      DIPImpModName( imp_image_handle *ii, imp_mod_handle im,
+                                char *buff, size_t buff_size );
